requirements.txt
better-profanity
aiohttp
discord-py
mysql-connector-python
psutil
flask
flask-discord